+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/service/register/ShenyuClientRegisterDefaultServiceImpl.java
@@ -32,7 +32,6 @@ import org.apache.shenyu.common.enums.DataEventTypeEnum;
 import org.apache.shenyu.common.enums.PluginEnum;
 import org.apache.shenyu.common.utils.GsonUtils;
 import org.apache.shenyu.register.common.dto.MetaDataRegisterDTO;
-import org.apache.shenyu.register.common.dto.URIRegisterDTO;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.stereotype.Service;
 
@@ -67,21 +66,7 @@ public class ShenyuClientRegisterDefaultServiceImpl extends AbstractShenyuClient
     }
 
     @Override
-    public String registerURI(final String contextPath, final List<String> uriList) {
-        SelectorDO selector = selectorService.findByName(contextPath);
-        SelectorData selectorData = selectorService.buildByName(contextPath);
-        String handler = GsonUtils.getInstance().toJson(buildDivideUpstreamList(uriList));
-        selector.setHandle(handler);
-        selectorData.setHandle(handler);
-        selectorService.updateSelective(selector);
-        // publish change event.
-        eventPublisher.publishEvent(new DataChangedEvent(ConfigGroupEnum.SELECTOR, DataEventTypeEnum.UPDATE,
-                Collections.singletonList(selectorData)));
-        return ShenyuResultMessage.SUCCESS;
-    }
-
-    @Override
-    public String registerURIDefault(final String contextPath, final List<URIRegisterDTO> registerDTOList) {
+    public String registerUri(final String contextPath, final List<String> uriList) {
         SelectorDO selector = selectorService.findByName(contextPath);
         SelectorData selectorData = selectorService.buildByName(contextPath);
         PluginDO pluginDO = pluginMapper.selectById(selector.getPluginId());
@@ -90,11 +75,11 @@ public class ShenyuClientRegisterDefaultServiceImpl extends AbstractShenyuClient
             if (Objects.nonNull(selector.getHandle())) {
                 SpringCloudSelectorHandle springCloudSelectorHandle = GsonUtils.getInstance()
                         .fromJson(selector.getHandle(), SpringCloudSelectorHandle.class);
-                springCloudSelectorHandle.setDivideUpstreams(buildDivideUpstreams(registerDTOList));
+                springCloudSelectorHandle.setDivideUpstreams(buildDivideUpstreamList(uriList));
                 handler = GsonUtils.getInstance().toJson(springCloudSelectorHandle);
             }
         } else {
-            handler = GsonUtils.getInstance().toJson(buildDivideUpstreams(registerDTOList));
+            handler = GsonUtils.getInstance().toJson(buildDivideUpstreamList(uriList));
         }
         selector.setHandle(handler);
         selectorData.setHandle(handler);
@@ -122,10 +107,5 @@ public class ShenyuClientRegisterDefaultServiceImpl extends AbstractShenyuClient
 
     private List<DivideUpstream> buildDivideUpstreamList(final List<String> uriList) {
         return uriList.stream().map(this::buildDivideUpstream).collect(Collectors.toList());
-    }
-
-    private List<DivideUpstream> buildDivideUpstreams(final List<URIRegisterDTO> registerDTOList) {
-        return registerDTOList.stream()
-                .map(s -> buildDivideUpstream(String.join(":", s.getHost(), s.getPort().toString()))).collect(Collectors.toList());
     }
 }